refactor: remove /home route in favor of the root path for landing page and homepage navigation
